Изменение MainActivity.kt и добавление кастомного EquilateralImageView.kt для отображения квадратных фотографий.
