--- app/src/main/java/ru/geekbrains/mymaterialproject/MainActivity.kt
@@ -1,18 +1,0 @@
-package ru.geekbrains.mymaterialproject
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import ru.geekbrains.mymaterialproject.ui.main.MainFragment
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.main_activity)
-        if (savedInstanceState == null) {
-            supportFragmentManager.beginTransaction()
-                .replace(R.id.container, MainFragment.newInstance())
-                .commitNow()
-        }
-    }
-}
